Squashed commit of the following:
commit e97daeca2f3073f1ccef2987991a3f07f53bb3ec
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Mon May 25 11:58:49 2026 +0100

    more stable PID compute

commit 6530121dbd15a2825af7c6229068affd6c0f1c04
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Tue May 19 17:02:43 2026 +0100

    bug fix

commit 1f32e393cc0ad2470bb084097b41e18ff8b80756
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Tue May 19 16:58:44 2026 +0100

    pid fixes and spindle range interger only math

commit fc8c2563410653c88d6111c74dbca21e6db38a7e
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Tue May 19 11:57:39 2026 +0100

    auto tunning for testing

    added auto tunning for testing

commit 1939fed950bf08cb9e5b39a64dce64ec0d90f002
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Tue May 19 11:28:55 2026 +0100

    modified PID and tool set_speed

    - fixed PID calculations time sample and absolute units
    - muted tool set speed to allow PID to full control the tool output
    - added Q15 version

commit af6180284d2ec93f6e890f9df52f8549d433949a
Merge: 0141831a 3eee3892
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Sat May 16 11:40:14 2026 +0100

    Merge branch 'master' into fixed-pid-tool-tests

commit 0141831a2b6c1b1bf39cdb985a079aba9e63766b
Author: Paciente8159 <paciente8159@gmail.com>
Date:   Fri May 15 17:22:39 2026 +0100

    Fixed PID tool testings

+++ uCNC/src/hal/tools/tool.c
@@ -253,7 +253,7 @@ int16_t tool_get_setpoint(void)
 {
 	// input value will always be positive
 #if TOOL_COUNT > 0
-	return tool_range_speed(tool_current_speed, 1);
+	return tool_range_speed(tool_current_speed, (uint8_t)1);
 #endif
 	return 0;
 }
